Rename TableWrapper to PaginatedLoader
